Moved icon types to tyoes dir
