+++ .editorconfig
@@ -22,6 +22,9 @@ file_header_template = Copyright (c) StoneyDSP.\nLicensed under the MIT License.
 indent_style = space
 indent_size = 2
 
+[*.css]
+indent_size = 2
+
 [*.{js,jsx,ts,tsx,mjs,cjs}]
 indent_style = space
 indent_size = 2
