--- components/icons/icons.d.ts
@@ -1,7 +1,0 @@
-export type IconSVG = {
-  title: string,
-  subtitle: string | undefined,
-  path: string,
-  viewBox: string,
-  url: string | undefined,
-}[]
